+++ analyzer/src/test/java/nl/basjes/parse/useragent/clienthints/TestClientHintsAnalysis.java
@@ -256,7 +256,7 @@ class TestClientHintsAnalysis {
         // ------------------------------------------
         // Without ClientHints
         testCase.expect("DeviceClass",                     "Phone");
-        testCase.expect("DeviceName",                      "Unknown");
+        testCase.expect("DeviceName",                      "Android Mobile");
         testCase.expect("DeviceBrand",                     "Unknown");
         testCase.expect("OperatingSystemClass",            "Mobile");
         testCase.expect("OperatingSystemName",             "Android");