+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This is intended as an overview of the major changes
 NEXT RELEASE
 ===
 - New/improved detections:
-  - ...
+  - Report the DeviceName "Android Mobile" for reduced Android user agents.
 
 v7.19.1
 ===